==> code/3_fer2013_csv_data_cleaning.ipynb <==
# Data Cleaning of .csv Data

### Imports

import pandas as pd
import warnings
warnings.filterwarnings("ignore", category=UserWarning)

# Reading in the .csv data
fer2013_data = pd.read_csv("../data/fer2013.csv")

fer2013_data.head()

# checking to see the split of the dataset between training and testing
fer2013_data['Usage'].value_counts()

# reshaping the dataset so it only contains one set of training and testing sets
fer2013_data = fer2013_data[fer2013_data['Usage'] != 'PrivateTest']

fer2013_data['Usage'].value_counts()

# splitting up the training data from the broader dataset
train_data = fer2013_data[fer2013_data['Usage'] == 'Training']

train_data.head(10)

# dropping the unnecessary usage column at this point
train_data.drop(columns = ['Usage'], inplace = True)

train_data

# checking the distribution of emotion classes within the training set
train_data['emotion'].value_counts(normalize = True)

fer2013_data['Usage'].value_counts()

# labeling the testing set from the broader dataset
test_data = fer2013_data[fer2013_data['Usage'] == 'PublicTest']

test_data.head()

# dropping usage column from testing set
test_data.drop(columns = 'Usage', inplace = True)

test_data.head()

train_data.to_csv('train_data.csv', index = False)

test_data.to_csv('test_data.csv', index = False)
